Add files via upload
Exercise 1 : There are 2 projects, a.) Aggregate is the Server that returns Random strings (Thought of the day messages). b.) Client makes async call to the API three times and calculates aggregate of the string length. If Cancel button is clicked the call is interrupted and a cancellation message is prompted.
Exercise 2 : Database considered for this exercise is MS SQL Server. 
Exercise 3 : Unit Test project for Coupon Manager project
Exercise 4 : A word document with code review comments highlights in bold Yellow.